Extract subschema to dedicated interface
