+++ packages/framework/src/Contracts/FrontMatter/PageSchema.php
@@ -5,18 +5,11 @@ namespace Hyde\Framework\Contracts\FrontMatter;
 /**
  * @see \Hyde\Framework\Concerns\HydePage
  */
-interface PageSchema
+interface PageSchema extends Support\NavigationSchema
 {
     public const PAGE_SCHEMA = [
         'title'         => 'string',
         'canonicalUrl'  => 'string|url',
         'navigation'    => 'array|navigation',
-    ];
-
-    public const NAVIGATION_SCHEMA = [
-        'label'     => 'string',
-        'group'     => 'string',
-        'hidden'    => 'bool',
-        'priority'  => 'int',
     ];
 }
