+++ packages/framework/src/Contracts/FrontMatter/Support/NavigationSchema.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Hyde\Framework\Contracts\FrontMatter\Support;
+
+interface NavigationSchema
+{
+    public const NAVIGATION_SCHEMA = [
+        'label'     => 'string',
+        'group'     => 'string',
+        'hidden'    => 'bool',
+        'priority'  => 'int',
+    ];
+}
